+++ src/main/gen/de/magynhard/crystal/psi/impl/CrystalPercentLiteralImpl.java
@@ -8,12 +8,11 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static de.magynhard.crystal.psi.CrystalTypes.*;
-import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import de.magynhard.crystal.psi.*;
 
-public class CrystalPercentLiteralImpl extends ASTWrapperPsiElement implements CrystalPercentLiteral {
+public class CrystalPercentLiteralImpl extends CrystalPercentLiteralMixin implements CrystalPercentLiteral {
 
-  public CrystalPercentLiteralImpl(@NotNull ASTNode node) {
+  public CrystalPercentLiteralImpl(ASTNode node) {
     super(node);
   }